Demo bug with second check on search job status
Summary: `Response::IdentitySearchJob#get` builds a new
`Command::RetrieveIdentitySearchJob` to check on the status of an async search
using the `Location` header. The `Location` header is not present when the
`Response::IdentitySearchJob` is the result of a previous
`Command::RetrieveIdentitySearchJob`.

Bug details: The sequence of events is documented by Cognito on this page:
https://cognitohq.com/docs/verifying-with-phone

When the `Command::CreateIdentitySearch` receives a 202 response, it will build
an initial `Response::IdentitySearchJob` and the underlying response contains
the `Location` header.

`Response::IdentitySearchJob#get` then builds an initial
`Command::RetrieveIdentitySearchJob` using the `Location` header.

If this initial command then recieves a 200 response with the status field set
to `processing`, the command will build a second `Response::IdentitySearchJob`
and the underlying response no longer contains the `Location` header.

Attempting to use the second `Response::IdentitySearchJob#get` method to build
a second `Command::RetrieveIdentitySearchJob` to check on the status of the
async search job will fail.